Split Interface_EditView into multiple files
The ruler and text objects were getting a bit big, so they have been split into
multiple files.
